DEVOPS-2195 add snyk scan
This is a return to the attempt to add snyk testing to CI. The issue was
with creating the volume in which to output the rendered template which
we then use to run snyk iac test against.

CircleCI didn't lke the `${GITHUB_WORKSPACE}` volume syntax; hopefully
this other syntax (`${{ github.workspace`) will work.