DEVOPS-2414 fix serviceAccount
I thought that the serviceAccount logic was broken. It wasn't; it's just
that it requires an `enabled: true` parameter that I wasn't aware of.

However, there was a logic error in that the serviceaccount template
expected to get its data from the global context, but the template is
always called with a specific serviceAccount context, and the global
assumption could be the source of bugs.

I'm still not sure I like the design of requiring an `enabled` flag, but
this is a simpler fix.

Now that I understand how this works, I've enabled a skipped test that
wasn't working because it was not referencing the correct file to
compare but also because I didn't know about the `enabled` flag. I've
also added testing to the Github Actions CI job.