DEVOPS-2657 Imperva support
Setting `imperva: true` on an Ingress will now change the behvior of the
resources created by the template. It will prepend `origin-` to any
hostnames (either defined or derived) and also add the
`alb.ingress.kubernetes.io/conditions` annotation with the original
hostname.

That annotation is defined here:
https://kubernetes-sigs.github.io/aws-load-balancer-controller/v2.7/guide/ingress/annotations/#traffic-routing

It sets a listener rule that sets the original hostname as a host header
for which it allows requests. Our `origin-` hostname will work with
external-dns and the AWS load balancer controller to route requests to
the EKS workload, and Imperva will route traffic from the web to the
original hostname via the `origin-` hostname. But it will send the
request with the host header containing the original hostname.

This hides all that detail behind a simple boolean.