Add option for services to redirect a base domain to www. for ALB
Adds a redirectHostToWWW option which takes the base domain to apply a
www redirect to and creates the relevant ALB annotations.

This option seems a little overly specific and like it could be a map
allowing arbitrary redirects to any of a service's hostnames, but went
with this since our only existing use case is for redirecting
sevenfifty.com domains to www and to avoid adding a lot of extra go
template code that might not be used.