INFRASEC-4249 make apiVersion a variable
The main purpose of this is to create an `apiVersion` variable for the
`secrets` section, defaulting to `v1beta1`, so that we can have
different versions of `external-secrets` running in different clusters.

The remaining changes are due to a change in Helm which doesn't deal
well with the symlinks we were using for the Helm Charts in the fixtures
dirs. This adds a script to replace the Charts and update the locks
whenever we update the main common chart.